\name{RTop500-package}
\alias{RTop500-package}
\docType{package}
\title{RTop500}
\description{
  Neither the package author nor the package itself has any 
  affiliation with the Top500 project.
  
  This software is licensed under the permissive 2-clause BSD license.
  You can find a quick summary of the license here:
  
  \url{https://tldrlegal.com/license/bsd-2-clause-license-\%28freebsd\%29}
}
\details{
\tabular{ll}{
  Package: \tab RTop500 \cr
  Type: \tab Package \cr
  License: \tab BSD 2-Clause \cr
}
}
\reference{
  Official Top500 website: \url{http://www.top500.org/}
  
  Raw Top500 data mirror: 
  \url{https://github.com/wrathematics/RTop500/tree/master/raw}
}
\author{
  Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
