\name{plot_top500}
\alias{plot_top500}


\title{Top500 Data Plotter}
\description{
  
}
\usage{
  plot_top500(type="theoretical", columns=c("Min", "Mean", "Max"), my.val=NULL)
}
\arguments{
  \item{columns}{A }
  \item{type}{}
  \item{my.val}{An optional value (matching the type argument) to add
  to the plot.  Passing NULL does nothing.}
}
\details{
  
}

