% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_top500}
\alias{plot_top500}
\title{Top500 Data Plotter}
\usage{
plot_top500(type = "theoretical", columns = c("Min", "Mean", "Max"),
  my.val = NULL)
}
\arguments{
\item{type}{A string indicating the type of quantity to plot.  Options are
"theoretical", "linpack", "difference", and "ncores".}

\item{columns}{A vector of strings that determines which values will be
shown on the graph.  Options include "Min", "Mean, "Max", "Median", and "Sum".}

\item{my.val}{An optional value (matching the type argument) to add to the
plot.  Want to see how your macine's performance compares with all
supercomputers since 1993?  Pass your own value and see.  Passing NULL does
nothing.}
}
\description{
\code{plot_top500} is a simple interface for being able to quickly summarize
historical supercomputing data.  It also includes an option for comparing
your own data against the historical supercomputing data.
}
\details{
TODO what do the options mean

Whichever combination of columns is passed will be plotted, and more than
one can be used.  For example, passing \code{columns=c("Min", "Max")} will
show the min and max lines.
}

