% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTop500-package.R
\name{RTop500-package}
\alias{RTop500-package}
\title{RTop500}
\description{
Neither the package author nor the package itself has any affiliation with
the Top500 project.
}
\details{
\tabular{ll}{ 
   Package: \tab RTop500 \cr 
   Type: \tab Package \cr 
   License: \tab BSD 2-Clause \cr 
}
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\references{
Official Top500 website: \url{http://www.top500.org/}

Raw Top500 data mirror:
\url{https://github.com/wrathematics/RTop500/tree/master/raw}
}
\keyword{Package}

