% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysfile.r
\name{RTop500_shinydir}
\alias{RTop500_shinydir}
\title{Location to Shiny Dir}
\usage{
RTop500_shinydir()
}
\description{
File path to shiny app for the RTop500 package.
}
\examples{
\dontrun{
library(RTop500)
shiny::runApp(RTop500_shinydir())
}

}

