\name{rgutenberg-package}
\alias{rgutenberg-package}
\docType{package}
\title{A Package for Scraping a Random Project Gutenberg Book}
\description{
  This R package downloads a random Project Gutenberg book.  The 
  purpose of this package is to offer for text, via the rgutenberg()
  function, somewhat similar functionality to that of rnorm(), runif(), 
  etc.  This functionality is really meant for developers who need
  quick access to random corpora for developing and testing methods
  that operate on text.  It is not meant to be used as an automated
  way of downloading large portions of the Project Gutenberg corpora;
  that would be insane.  Please also be considerate to their servers
  and don't abuse this functionality.
  
  Neither the package author, nor the package itself, has any 
  affiliation with Project Gutenberg.
  
  This software is licensed under the permissive 2-clause BSD license.
  You can find a quick summary of the license here:
  
  \url{https://tldrlegal.com/license/bsd-2-clause-license-\%28freebsd\%29}
  
  The full terms of the license (it's very short) are contained in the
  LICENSE file in the root directory of the project.
}
\details{
\tabular{ll}{
  Package: \tab rgutenberg\cr
  Type: \tab Package\cr
  License: \tab BSD 2-Clause\cr
}
  This package requires an active internet connection to use.
}
\references{
  Project Gutenberg: \url{http://gutenberg.org/}
}
\author{
  Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
